API

API

Types

ArraysOfArrays.AbstractArrayOfSimilarArrays

ArraysOfArrays.ArrayOfSimilarArrays

ArraysOfArrays.VectorOfArrays

Functions

ArraysOfArrays.deepmap

ArraysOfArrays.flatview

ArraysOfArrays.flatview

ArraysOfArrays.flatview

ArraysOfArrays.innersize

ArraysOfArrays.nestedmap2

ArraysOfArrays.nestedview

Base.sum

Statistics.cor

Statistics.cov

Statistics.mean

Statistics.var

Documentation

ArraysOfArrays.AbstractArrayOfSimilarArrays — Type.

AbstractArrayOfSimilarArrays{T,M,N} <: AbstractArray{<:AbstractArray{T,M},N}

An array that contains arrays that have the same size/axes. The array is internally stored in flattened form as some kind of array of dimension M + N. The flattened form can be accessed via flatview(A).

Subtypes must implement (in addition to typical array operations):

flatview(A::SomeArrayOfSimilarArrays)::AbstractArray{T,M+N}

The following type aliases are defined:

AbstractVectorOfSimilarArrays{T,M} = AbstractArrayOfSimilarArrays{T,M,1}

AbstractArrayOfSimilarVectors{T,N} = AbstractArrayOfSimilarArrays{T,1,N}

AbstractVectorOfSimilarVectors{T} = AbstractArrayOfSimilarArrays{T,1,1}

source

ArraysOfArrays.ArrayOfSimilarArrays — Type.

ArrayOfSimilarArrays{T,M,N,L,P} <: AbstractArrayOfSimilarArrays{T,M,N}

Represents a view of an array of dimension L = M + N as an array of dimension M with elements that are arrays with dimension N. All element arrays implicitly have equal size/axes.

Constructors:

ArrayOfSimilarArrays{T,M,N}(flat_data::AbstractArray)
ArrayOfSimilarArrays{T,M}(flat_data::AbstractArray)

The following type aliases are defined:

VectorOfSimilarArrays{T,M} = AbstractArrayOfSimilarArrays{T,M,1}

ArrayOfSimilarVectors{T,N} = AbstractArrayOfSimilarArrays{T,1,N}

VectorOfSimilarVectors{T} = AbstractArrayOfSimilarArrays{T,1,1}

VectorOfSimilarArrays supports push!(), etc., provided the underlying array supports resizing of it's last dimension (e.g. an ElasticArray).

The nested array can also be created using the function nestedview and the wrapped flat array can be accessed using flatview afterwards:

A_flat = rand(2,3,4,5,6)
A_nested = nestedview(A_flat, 2)
A_nested isa AbstractArray{<:AbstractArray{T,2},3} where T
flatview(A_nested) === A_flat

source

ArraysOfArrays.VectorOfArrays — Type.

VectorOfArrays{T,N,M} <: AbstractVector{<:AbstractArray{T,N}}

An VectorOfArrays represents a vector of N-dimensional arrays (that may differ in size). Internally, VectorOfArrays stores all elements of all arrays in a single flat vector. M must equal N - 1

The VectorOfArrays itself supports push!, unshift!, etc., but the size of each individual array in the vector is fixed. resize! can be used to shrink, but not to grow, as the size of the additional element arrays in the vector would be unknown. However, memory space for up to n arrays with a maximum size s can be reserved via sizehint!(A::VectorOfArrays, n, s::Dims{N}).

Constructors:

VectorOfArrays{T, N}()

VectorOfArrays(
    data::AbstractVector,
    elem_ptr::AbstractVector{Int},
    kernel_size::AbstractVector{<:Dims}
    checks::Function = ArraysOfArrays.full_consistency_checks
)

Other suitable values for checks are ArraysOfArrays.simple_consistency_checks and ArraysOfArrays.no_consistency_checks.

The following type aliases are defined:

VectorOfVectors{T,VT,VI,VD} = VectorOfArrays{T,1,VT,VI,VD}

source

ArraysOfArrays.deepmap — Function.

deepmap(f::Base.Callable, x::Any)
deepmap(f::Base.Callable, A::AbstractArray{<:AbstractArray{<:...}})

Applies map at the deepest possible layer of nested arrays. If A is not a nested array, deepmap behaves identical to Base.map.

source

ArraysOfArrays.flatview — Function.

flatview(A::AbstractArray)
flatview(A::AbstractArray{<:AbstractArray{<:...}})

View array A in a suitable flattened form. The shape of the flattened form will depend on the type of A. If the A is not a nested array, the return value is A itself. When no type-specific method is available, flatview will fall back to Base.Iterators.flatten.

source

ArraysOfArrays.flatview — Method.

flatview(A::ArrayOfSimilarArrays{T,M,N,L,P})::P

Returns the array of dimensionality L = M + N wrapped by A. The shape of the result may be freely changed without breaking the inner consistency of A.

source

ArraysOfArrays.flatview — Method.

flatview(A::VectorOfArrays{T})::Vector{T}

Returns the internal serialized representation of all element arrays of A as a single vector. Do not change the length of the returned vector, as it would break the inner consistency of A.

source

ArraysOfArrays.innersize — Function.

innersize(A:AbstractArray{<:AbstractArray}, [dim])

Returns the size of the element arrays of A. Fails if the element arrays are not of equal size.

source

ArraysOfArrays.nestedmap2 — Function.

nestedmap2(f::Base.Callable, A::AbstractArray{<:AbstractArray})

Nested map at depth 2. Equivalent to map(X -> map(f, X) A).

source

ArraysOfArrays.nestedview — Function.

nestedview(A::AbstractArray{T,M+N}, M::Integer)
nestedview(A::AbstractArray{T,2})

AbstractArray{<:AbstractArray{T,M},N}

View array A in as an M-dimensional array of N-dimensional arrays by wrapping it into an ArrayOfSimilarArrays.

source

Base.sum — Method.

sum(X::AbstractVectorOfSimilarArrays)
sum(X::AbstractVectorOfSimilarArrays, w::StatsBase.AbstractWeights)

Compute the sum of the elements vectors of X. Equivalent to sum of flatview(X) along the last dimension.

source

Statistics.cor — Method.

cor(X::AbstractVectorOfSimilarVectors)
cor(X::AbstractVectorOfSimilarVectors, w::StatsBase.AbstractWeights)

Compute the Pearson correlation matrix between the elements of the elements of X along X. Equivalent to cor of flatview(X) along dimension 2.

source

Statistics.cov — Method.

cov(X::AbstractVectorOfSimilarVectors; corrected::Bool = true)
cov(X::AbstractVectorOfSimilarVectors, w::StatsBase.AbstractWeights; corrected::Bool = true)

Compute the covariance matrix between the elements of the elements of X along X. Equivalent to cov of flatview(X) along dimension 2.

source

Statistics.mean — Method.

mean(X::AbstractVectorOfSimilarArrays)
mean(X::AbstractVectorOfSimilarArrays, w::StatsBase.AbstractWeights)

Compute the mean of the elements vectors of X. Equivalent to mean of flatview(X) along the last dimension.

source

Statistics.var — Method.

var(X::AbstractVectorOfSimilarArrays; corrected::Bool = true)
var(X::AbstractVectorOfSimilarArrays, w::StatsBase.AbstractWeights; corrected::Bool = true)

Compute the sample variance of the elements vectors of X. Equivalent to var of flatview(X) along the last dimension.

source